Incluir expansão de perfil por módulos e capacidades

+++ back-end/database/seeders/TipoCapacidadeSeeder.php
@@ -6,6 +6,7 @@ use Illuminate\Database\Seeder;
 use App\Models\TipoCapacidade;
 use App\Models\Capacidade;
 use App\Models\Perfil;
+use Ramsey\Uuid\Uuid;
 use App\Services\TipoCapacidadeService;
 use App\Services\UtilService;
 
@@ -18,6 +19,7 @@ class TipoCapacidadeSeeder extends Seeder
      */
     public function run()
     {
+        //TipoCapacidade::whereIn('codigo',['MOD_LOGS_CONS','MOD_LOGS_EDT','MOD_LOGS_EXCL'])->delete(); //apaga capacidades inseridas indevidamente
         $tiposCapacidadesService = new TipoCapacidadeService();
         $utilService = new UtilService();
 
@@ -37,21 +39,47 @@ class TipoCapacidadeSeeder extends Seeder
         $developerId = $perfil->id;
 
         // carrega os tipos de capacidades do vetor declarado no serviço TipoCapacidadeService
-        $dadosTiposCapacidades = array_map(fn ($capacidade) => array_merge([$utilService->uuid($capacidade[0])], $capacidade), $tiposCapacidadesService->tiposCapacidades);
+        $dadosTiposCapacidades = array_map(fn ($capacidade) => array_merge([$utilService->uuid($capacidade['codigo'])], $capacidade), $tiposCapacidadesService->tiposCapacidades);
         foreach ($dadosTiposCapacidades as $registro) {
             //insere no banco ou atualiza todos os tipos de capacidade constantes do vetor
             $tipocapacidade = TipoCapacidade::where('id', $registro[0])->first() ?? new TipoCapacidade();
             $tipocapacidade->fill([
                 'id' => $registro[0],
-                'codigo' => $registro[1],
-                'descricao' => $registro[2]
+                'codigo' => $registro['codigo'],
+                'descricao' => $registro['descricao'],
+                'grupo_id' => NULL
             ]);
             $tipocapacidade->save();
-            //garante que o perfil de Desenvolver tenha todos os tipos de permissão
+            $dadosCapacidadesFilhas = array_map(fn ($capacidadeFilha) => array_merge([$utilService->uuid($capacidadeFilha[0])], $capacidadeFilha), $registro['capacidades']);
+            foreach ($dadosCapacidadesFilhas as $registroFilha) {
+                $tipocapacidadeFilha = TipoCapacidade::where('id', $registroFilha[0])->first() ?? new TipoCapacidade();
+                $tipocapacidadeFilha->fill([
+                    'id' => $registroFilha[0],
+                    'codigo' => $registroFilha[1],
+                    'descricao' => $registroFilha[2],
+                    'grupo_id' => $registro[0]
+                ]);
+                $tipocapacidadeFilha->save();
+                //garante que o perfil de Desenvolvedor tenha todos os tipos de capacidades filhas
+                if (!Capacidade::where([['perfil_id', $developerId], ['tipo_capacidade_id', $registroFilha[0]]])->exists()) {
+                    $capacidade = new Capacidade();
+                    $capacidade->fill([
+                        'id' => $utilService->uuid(),
+                        'data_inicio' => date("Y-m-d H:i:s"),
+                        'data_fim' => null,
+                        'perfil_id' => $developerId,
+                        'tipo_capacidade_id' => $registroFilha[0]
+                    ]);
+                    $capacidade->save();
+            }
+            }
+            //garante que o perfil de Desenvolvedor tenha todos os tipos de permissão
             if (!Capacidade::where([['perfil_id', $developerId], ['tipo_capacidade_id', $registro[0]]])->exists()) {
                 $capacidade = new Capacidade();
                 $capacidade->fill([
                     'id' => $utilService->uuid(),
+                    'data_inicio' => date("Y-m-d H:i:s"),
+                    'data_fim' => null,
                     'perfil_id' => $developerId,
                     'tipo_capacidade_id' => $registro[0]
                 ]);
@@ -60,10 +88,23 @@ class TipoCapacidadeSeeder extends Seeder
         }
 
         // exclui as capacidades que não existem mais no vetor declarado no serviço TipoCapacidadeService
-        foreach (TipoCapacidade::whereNotIn('codigo',array_map(fn($x) => $x[0],$tiposCapacidadesService->tiposCapacidades))->get() as $tipo) {
-            if(count($tipo->capacidades) > 0) Capacidade::where('tipo_capacidade_id',$tipo->id)->delete();
-            $tipo->delete();
+        $dadosModulosCapacidades = array_map(fn($x) => $x,$tiposCapacidadesService->tiposCapacidades);
+        
+        foreach ($dadosModulosCapacidades as $modulo) {
+            //foreach ($modulo['capacidades'] as $dadosModulosCapacidadesFilhas) {
+            $arrayCapacidades = array_map(fn($z) => $z[0], $modulo['capacidades']);
+            $tiposCapacidadesNulos = TipoCapacidade::where('grupo_id', $utilService->uuid($modulo['codigo']))
+                                                    ->whereNotIn('codigo',array_map(fn($z) => $z[0], $modulo['capacidades']))->get();
+            $i = 1;
+            foreach (TipoCapacidade::where('grupo_id', $utilService->uuid($modulo['codigo']))
+                                    ->whereNotIn('codigo',array_map(fn($z) => $z[0], $modulo['capacidades']))->get() as $tipo) {
+                foreach($tipo->capacidades as $tipoCap) {
+                    $cap = Capacidade::where('tipo_capacidade_id',$tipo->id)->first();
+                    $cap->delete();               
+                }
+                if ($tipo->grupo_id) $tipo->delete();
+                $i++;
+            }
         }
-
     }
 }
